+++ packages/type-extrinsics/src/fromMetadata/createUnchecked.ts
@@ -34,7 +34,7 @@ export default function createDescriptor (
     return new Method({
       args,
       callIndex
-    }, callMetadata);
+    }, { meta: callMetadata });
   };
 
   extrinsicFn.callIndex = callIndex;